# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"","version":"0.0.1"},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/categories/{from}/{to}":{"get":{"tags":["get by categories"],"summary":"getAggregatedScores get by categories","description":"Get aggregated scores for categories within a specified period.","operationId":"get by categories#getAggregatedScores","parameters":[{"name":"from","in":"path","description":"Start date in YYYY-MM-DD format","required":true,"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"name":"to","in":"path","description":"End date in YYYY-MM-DD format","required":true,"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/GetByCategoriesGetAggregatedScoresResponseBody","required":["meta","data"]}}},"schemes":["http"]}},"/overall/{from}/{to}":{"get":{"tags":["get overall"],"summary":"getOverallScore get overall","description":"Get the overall aggregate score for a specified period.","operationId":"get overall#getOverallScore","parameters":[{"name":"from","in":"path","description":"Start date in YYYY-MM-DD format","required":true,"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"name":"to","in":"path","description":"End date in YYYY-MM-DD format","required":true,"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"}],"responses":{"200":{"description":"OK response.","schema":{"type":"number","format":"float"}}},"schemes":["http"]}},"/periods/{period}":{"get":{"tags":["get changes by periods"],"summary":"getChangesByPeriods get changes by periods","description":"Get the score change from a selected period over the previous period.","operationId":"get changes by periods#getChangesByPeriods","parameters":[{"name":"period","in":"path","description":"The period type (e.g., week, month, year)","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/GetChangesByPeriodsGetChangesByPeriodsResponseBody","required":["meta","data"]}}},"schemes":["http"]}},"/tickets/{from}/{to}":{"get":{"tags":["get by tickets"],"summary":"getAggregatedScoresByTicket get by tickets","description":"Get aggregate category scores by ticket for a specified period.","operationId":"get by tickets#getAggregatedScoresByTicket","parameters":[{"name":"from","in":"path","description":"Start date in YYYY-MM-DD format","required":true,"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"name":"to","in":"path","description":"End date in YYYY-MM-DD format","required":true,"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/GetByTicketsGetAggregatedScoresByTicketResponseBody","required":["meta","data"]}}},"schemes":["http"]}}},"definitions":{"CategoryMetaResponseBody":{"title":"CategoryMetaResponseBody","type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/definitions/CategoryResponseBody"},"example":[{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."}]},"periods":{"type":"array","items":{"$ref":"#/definitions/PeriodResponseBody"},"example":[{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073}]}},"example":{"categories":[{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."}],"periods":[{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073}]}},"CategoryResponseBody":{"title":"CategoryResponseBody","type":"object","properties":{"id":{"type":"integer","example":5993102946133411412,"format":"int64"},"name":{"type":"string","example":"Rerum qui provident quas."}},"example":{"id":3315688220268315851,"name":"Quo ut nostrum eaque asperiores."}},"CategoryScoreDetailResponseBody":{"title":"CategoryScoreDetailResponseBody","type":"object","properties":{"id":{"type":"integer","example":574433862816017834,"format":"int64"},"score":{"type":"number","example":0.4649607,"format":"float"}},"example":{"id":5031206183273705840,"score":0.5823075}},"CategoryScoreResponseBody":{"title":"CategoryScoreResponseBody","type":"object","properties":{"category_id":{"type":"integer","example":2068249123307792795,"format":"int64"},"num_of_reviews":{"type":"integer","example":6974205976546608499,"format":"int64"},"periods":{"type":"array","items":{"$ref":"#/definitions/ScorePeriodResponseBody"},"example":[{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106}]},"total_score":{"type":"number","example":0.23293424,"format":"float"}},"example":{"category_id":812103680925109811,"num_of_reviews":7177905466422160222,"periods":[{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106}],"total_score":0.8073685}},"GetByCategoriesGetAggregatedScoresResponseBody":{"title":"GetByCategoriesGetAggregatedScoresResponseBody","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/CategoryScoreResponseBody"},"example":[{"category_id":6012716289030742144,"num_of_reviews":1693922704743648072,"periods":[{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106}],"total_score":0.6612983},{"category_id":6012716289030742144,"num_of_reviews":1693922704743648072,"periods":[{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106}],"total_score":0.6612983}]},"meta":{"$ref":"#/definitions/CategoryMetaResponseBody"}},"example":{"data":[{"category_id":6012716289030742144,"num_of_reviews":1693922704743648072,"periods":[{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106}],"total_score":0.6612983},{"category_id":6012716289030742144,"num_of_reviews":1693922704743648072,"periods":[{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106},{"id":8266584712360546909,"score":0.8505106}],"total_score":0.6612983}],"meta":{"categories":[{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."}],"periods":[{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073}]}},"required":["meta","data"]},"GetByTicketsGetAggregatedScoresByTicketResponseBody":{"title":"GetByTicketsGetAggregatedScoresByTicketResponseBody","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/TicketScoreResponseBody"},"example":[{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":5260201506769468096},{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":5260201506769468096},{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":5260201506769468096},{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":5260201506769468096}]},"meta":{"$ref":"#/definitions/TicketMetaResponseBody"}},"example":{"data":[{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":5260201506769468096},{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":5260201506769468096}],"meta":{"categories":[{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."}]}},"required":["meta","data"]},"GetChangesByPeriodsGetChangesByPeriodsResponseBody":{"title":"GetChangesByPeriodsGetChangesByPeriodsResponseBody","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/PeriodScoreDeltaResponseBody"},"example":[{"period_id":1665803925861920822,"score_delta":0.47400105},{"period_id":1665803925861920822,"score_delta":0.47400105},{"period_id":1665803925861920822,"score_delta":0.47400105}]},"meta":{"$ref":"#/definitions/PeriodsMetaResponseBody"}},"example":{"data":[{"period_id":1665803925861920822,"score_delta":0.47400105},{"period_id":1665803925861920822,"score_delta":0.47400105}],"meta":{"periods":[{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073}]}},"required":["meta","data"]},"PeriodResponseBody":{"title":"PeriodResponseBody","type":"object","properties":{"end":{"type":"integer","example":7740342830856434916,"format":"int64"},"id":{"type":"integer","example":8391435840159832747,"format":"int64"},"start":{"type":"integer","example":424218092375798371,"format":"int64"}},"example":{"end":991735833441535012,"id":8178566988592527494,"start":5806412509653981942}},"PeriodScoreDeltaResponseBody":{"title":"PeriodScoreDeltaResponseBody","type":"object","properties":{"period_id":{"type":"integer","example":1871308469996485659,"format":"int64"},"score_delta":{"type":"number","example":0.20171677,"format":"float"}},"example":{"period_id":7196622766114100189,"score_delta":0.0026400213}},"PeriodsMetaResponseBody":{"title":"PeriodsMetaResponseBody","type":"object","properties":{"periods":{"type":"array","items":{"$ref":"#/definitions/PeriodResponseBody"},"example":[{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073}]}},"example":{"periods":[{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073},{"end":8568520418160968230,"id":2293668282350108161,"start":5157908177095075073}]}},"ScorePeriodResponseBody":{"title":"ScorePeriodResponseBody","type":"object","properties":{"id":{"type":"integer","example":2278869036452526071,"format":"int64"},"score":{"type":"number","example":0.33896387,"format":"float"}},"example":{"id":5665419741085241523,"score":0.08853352}},"TicketMetaResponseBody":{"title":"TicketMetaResponseBody","type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/definitions/CategoryResponseBody"},"example":[{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."}]}},"example":{"categories":[{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."},{"id":5220825332390222813,"name":"Aut aut dignissimos."}]}},"TicketScoreResponseBody":{"title":"TicketScoreResponseBody","type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/definitions/CategoryScoreDetailResponseBody"},"example":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}]},"ticket_id":{"type":"integer","example":3930929717618624115,"format":"int64"}},"example":{"categories":[{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873},{"id":4316543418170283353,"score":0.23660873}],"ticket_id":4316262096533562179}}}}